# frontend/lib/openapi/BackendService.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LinkList Service",
    "description": "LinkList Service API to serve links and categories",
    "contact": {
      "name": "Andreas Lange",
      "email": "andreas@cfapps.cool"
    },
    "license": {
      "name": "MIT License",
      "url": "<https://opensource.org/license/mit/>"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "localhost dev"
    },
    {
      "url": "https://linklist.cfapps.cool",
      "description": "prod environment"
    }
  ],
...
  "paths": {
    "/api/v1/links/{id}": {
      "put": {
        "tags": [
          "link-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLinkDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LinkDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "link-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/links": {
      "get": {
        "tags": [
          "link-controller"
        ],
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "link-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LinkDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/links/byTitle/{title}": {
      "get": {
        "tags": [
          "link-controller"
        ],
        "operationId": "getByCategory",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/byId/{id}": {
      "get": {
        "tags": [
          "link-controller"
        ],
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LinkDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/byCategory/{id}": {
      "get": {
        "tags": [
          "link-controller"
        ],
        "operationId": "getByCategory_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateLinkDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "category": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "support": {
            "type": "boolean"
          },
          "contact": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "LinkDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "category": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "support": {
            "type": "boolean"
          },
          "contact": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "categoryTitle": {
            "type": "string"
          }
        }
      },
...
      "CreateLinkDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "support": {
            "type": "boolean"
          },
          "contact": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}